requirements.txt
tensorflow==2.16.1
fastapi
uvicorn
python-multipart
pillow
matplotlib
numpy